+++ src/atoms/Link.jsx
@@ -4,28 +4,36 @@ import PropTypes from 'prop-types';
 import * as colors from '../tokens/colors';
 import * as typography from '../tokens/typography';
 
-function selectLink(type) {
-  switch (type) {
-    case 'accent':
-      return colors.accent;
-    case 'disabled':
-      return colors.gray400;
-    case 'info':
-      return colors.info;
-    case 'error':
-      return colors.error;
-    default:
-      return '';
-  }
-}
+const colorTypes = {
+  accent: {
+    color: colors.accent,
+    hover: colors.accentStrong,
+  },
+  disabled: {
+    color: colors.gray400,
+    hover: colors.gray400,
+  },
+  info: {
+    color: colors.info,
+    hover: colors.infoStrong,
+  },
+  error: {
+    color: colors.error,
+    hover: colors.errorStrong,
+  },
+};
 
 const Link = ({ large, text, href, type }) => {
   const linkStyle = css`
     font-size: ${large ? `${typography.sizeL}` : `${typography.sizeS}`};
     font-weight: ${typography.semibold};
     text-decoration: underline;
-    color: ${selectLink(type)};
+    color: ${colorTypes[type].color};
     font-family: 'Open Sans', sans-serif;
+
+    &:hover {
+      color: ${colorTypes[type].hover};
+    }
   `;
 
   return (