+++ src/tokens/colors.js
@@ -14,4 +14,6 @@ export const gray100 = '#FAFAFA';
 export const success = '#66BA5B';
 export const warning = '#E8B600';
 export const error = '#FF040D';
+export const errorStrong = '#C80007';
 export const info = '#00ABCB';
+export const infoStrong = '#009AB7';